Moved display to sofast
