+++ example/fixed_pattern_deflectometry/calculate_dot_locations_from_display_object.py
@@ -7,7 +7,7 @@ from os.path import join, dirname, exists
 
 from opencsp.app.sofast.lib.DotLocationsFixedPattern import DotLocationsFixedPattern
 from opencsp.app.sofast.lib.SystemSofastFixed import SystemSofastFixed
-from opencsp.common.lib.deflectometry.Display import Display
+from opencsp.app.sofast.lib.Display import Display
 from opencsp.common.lib.deflectometry.ImageProjection import ImageProjection
 from opencsp.app.sofast.lib.SpatialOrientation import SpatialOrientation
 from opencsp.common.lib.opencsp_path.opencsp_root_path import opencsp_code_dir
