+++ opencsp/app/sofast/calibration/lib/save_physical_setup_file.py
@@ -3,7 +3,7 @@
 from numpy import ndarray
 from scipy.spatial.transform import Rotation
 
-from opencsp.common.lib.deflectometry.Display import Display
+from opencsp.app.sofast.lib.Display import Display
 from opencsp.common.lib.geometry.Vxy import Vxy
 from opencsp.common.lib.geometry.Vxyz import Vxyz
