+++ opencsp/app/sofast/lib/ProcessSofastFringe.py
@@ -13,7 +13,7 @@ from opencsp.common.lib.csp.Facet import Facet
 from opencsp.common.lib.csp.FacetEnsemble import FacetEnsemble
 from opencsp.common.lib.csp.MirrorPoint import MirrorPoint
 import opencsp.app.sofast.lib.calculation_data_classes as cdc
-from opencsp.common.lib.deflectometry.Display import Display
+from opencsp.app.sofast.lib.Display import Display
 from opencsp.common.lib.deflectometry.EnsembleData import EnsembleData
 from opencsp.common.lib.deflectometry.FacetData import FacetData
 import opencsp.app.sofast.lib.image_processing as ip
